internal_service/requirements.txt
flask
pyyaml==5.3.1
